GH-10965: Add module initializer functionality.
