+++ projects/core/src/base-core.module.ts
@@ -1,9 +1,4 @@
-import {
-  APP_INITIALIZER,
-  ModuleWithProviders,
-  NgModule,
-  Optional,
-} from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CmsModule } from './cms/cms.module';
 import { ConfigInitializerModule } from './config/config-initializer/config-initializer.module';
 import { ConfigValidatorModule } from './config/config-validator/config-validator.module';
@@ -11,31 +6,13 @@ import { ConfigModule } from './config/config.module';
 import { FeaturesConfigModule } from './features-config/features-config.module';
 import { GlobalMessageModule } from './global-message/global-message.module';
 import { I18nModule } from './i18n/i18n.module';
-import { LazyModulesService } from './lazy-loading/lazy-modules.service';
-import { MODULE_INITIALIZER } from './lazy-loading/tokens';
+import { LazyLoadingModule } from './lazy-loading/lazy-loading.module';
 import { BaseOccModule } from './occ/base-occ.module';
 import { MetaTagConfigModule } from './occ/config/meta-tag-config.module';
 import { ProcessModule } from './process/process.module';
 import { SiteContextModule } from './site-context/site-context.module';
 import { StateModule } from './state/state.module';
 
-export function moduleInitializersFactory(
-  lazyModuleService: LazyModulesService,
-  moduleInitializerFunctions: (() => any)[]
-): () => any {
-  return () => {
-    Promise.all(
-      lazyModuleService.runModuleInitializerFunctions(
-        moduleInitializerFunctions
-      )
-    ).catch((error) =>
-      console.error(
-        `MODULE_INITIALIZER promise was rejected when app was initialized: `,
-        error
-      )
-    );
-  };
-}
 @NgModule({
   imports: [
     StateModule.forRoot(),
@@ -50,20 +27,13 @@ export function moduleInitializersFactory(
     SiteContextModule.forRoot(), // should be imported after RouterModule.forRoot, because it overwrites UrlSerializer
     MetaTagConfigModule.forRoot(),
     BaseOccModule.forRoot(),
+    LazyLoadingModule.forRoot(),
   ],
 })
 export class BaseCoreModule {
   static forRoot(): ModuleWithProviders<BaseCoreModule> {
     return {
       ngModule: BaseCoreModule,
-      providers: [
-        {
-          provide: APP_INITIALIZER,
-          useFactory: moduleInitializersFactory,
-          deps: [LazyModulesService, [new Optional(), MODULE_INITIALIZER]],
-          multi: true,
-        },
-      ],
     };
   }
 }
