+++ projects/core/src/lazy-loading/lazy-loading.module.ts
@@ -1,0 +1,43 @@
+import {
+  APP_INITIALIZER,
+  ModuleWithProviders,
+  NgModule,
+  Optional,
+} from '@angular/core';
+import { LazyModulesService } from './lazy-modules.service';
+import { MODULE_INITIALIZER } from './tokens';
+
+export function moduleInitializersFactory(
+  lazyModuleService: LazyModulesService,
+  moduleInitializerFunctions: (() => any)[]
+): () => any {
+  return () => {
+    Promise.all(
+      lazyModuleService.runModuleInitializerFunctions(
+        moduleInitializerFunctions
+      )
+    ).catch((error) =>
+      console.error(
+        `MODULE_INITIALIZER promise was rejected when app was initialized: `,
+        error
+      )
+    );
+  };
+}
+
+@NgModule({})
+export class LazyLoadingModule {
+  static forRoot(): ModuleWithProviders<LazyLoadingModule> {
+    return {
+      ngModule: LazyLoadingModule,
+      providers: [
+        {
+          provide: APP_INITIALIZER,
+          useFactory: moduleInitializersFactory,
+          deps: [LazyModulesService, [new Optional(), MODULE_INITIALIZER]],
+          multi: true,
+        },
+      ],
+    };
+  }
+}
